+++ cli/__init__.py
@@ -1,0 +1,7 @@
+"""
+Command-Line Interface Module
+
+Provides CLI tools for issue analysis and duplicate detection.
+"""
+
+__version__ = "1.0.0"
