+++ ui/__init__.py
@@ -1,0 +1,7 @@
+"""
+User Interface Module
+
+Provides web-based UI for interactive issue analysis.
+"""
+
+__version__ = "1.0.0"
