+++ utils/duplicate/cosine_duplicate.py
@@ -7,7 +7,7 @@ import numpy as np
 from sklearn.feature_extraction.text import TfidfVectorizer
 from sklearn.metrics.pairwise import cosine_similarity
 
-from models import DuplicateDetectionResult, IssueReference
+from utils.models import DuplicateDetectionResult, IssueReference
 
 
 class CosineDuplicateAnalyzer:
