+++ utils/security/__init__.py
@@ -1,0 +1,9 @@
+"""
+Security Module
+
+Provides prompt injection detection and other security features.
+"""
+
+from utils.security.prompt_injection import PromptInjectionDetector
+
+__all__ = ["PromptInjectionDetector"]
