+++ utils/security/__main__.py
@@ -1,0 +1,6 @@
+"""Entry point for running prompt_injection as a module."""
+
+if __name__ == "__main__":
+    from utils.security.prompt_injection import main
+
+    main()
